+++ Parley/Parley/Program.cs
@@ -2,7 +2,7 @@
 using System;
 using System.IO;
 using System.Runtime.InteropServices;
-using Parley.Services;
+using DialogEditor.Services;
 
 namespace DialogEditor;